refactor: combine authoring apps to openedx_content
This commit makes the following major refactorings:

- All apps previously under openedx_learning/apps/authoring have now
  effectively been merged together to one openedx_content app.
- The models and logic for those apps continues to be encapsulated in
  what we're calling "applets", inside of openedx_content/applets/...
- In order to facilitate smooth database migrations, the AppConfigs and
  database migrations for the old apps continue to be preserved in
  openedx_content/backcompat/...
- A new openedx_learning.api.django.openedx_learning_apps_to_install()
  has been created to make it easier for openedx-platform to get all the
  necessary apps to install.

The rationale for this change is detailed in
docs/decisions/0020-authoring-as-one-app.rst, but the short version is
that we hope this arrangement will allow us to keep many of the benefits
of having small apps (easy to reason about), while also making it easier
to refactor internally. Refactoring is currently hindered by the
difficulty in moving models across apps.

The name change away from "authoring" also reflects that we intend to
use these APIs and models on the "learning" side of things as well, e.g.
when reading content for rendering to a student. In the longer term, we
will probably make openedx_content a top-level package in this repo, but
that would be a later step.

Most of this commit is just shuffling things around and renaming
references, but the truly complicated bits are around the sequencing of
database migrations, particularly the ones removing model state from the
old apps, and transferring that state into the new openedx_content app
without changing the actual database state. We also have some hacky
logic in openedx_content/apps.py in order to properly patch migration
dependencies so that we don't break existing openedx-platform
migrations.

This also bumps the version to 0.31.0.